requirements.txt
# ── Core HTTP & parsing ─────────────────────────────────────────────
httpx>=0.27.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
lxml>=5.2.0
python-dateutil>=2.9.0

# ── Data validation & config ────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# ── Templating & rendering ──────────────────────────────────────────
jinja2>=3.1.4

# ── Resilience ──────────────────────────────────────────────────────
tenacity>=8.3.0

# ── Testing ─────────────────────────────────────────────────────────
pytest>=8.2.0

# ── Optional: AI-generated Spanish summaries ────────────────────────
# openai>=1.30.0
